requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.3
Django==4.2.25
idna==3.11
psycopg2==2.9.11
pyserial==3.5
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
